:pencil: code review

+++ src/mc_outputs.rs
@@ -54,7 +54,7 @@ pub fn read_max_spendable_mc_index(db: &Connection, kind: &str) -> Result<u32> {
 pub struct McIndexInterval {
     pub from_mci: u32,
     pub to_mci: u32,
-    pub accumulated: u32,
+    pub accumulated: i64,
     pub has_sufficient: bool,
 }
 
@@ -86,7 +86,7 @@ pub fn find_mc_index_interval_to_target_amount(
 
     //Original js has another implementation for mysql
     let min_mc_output = if kind == "witnessing" { 11.0 } else { 344.0 };
-    let max_count_outputs = (target_amount as f64 / min_mc_output).ceil() as u32;
+    let max_count_outputs = (target_amount as f64 / min_mc_output).ceil() as i64;
 
     let sql = format!(
         "SELECT main_chain_index, amount \
@@ -98,7 +98,7 @@ pub fn find_mc_index_interval_to_target_amount(
 
     struct Row {
         main_chain_index: u32,
-        amount: u32,
+        amount: i64,
     }
 
     let mut stmt = db.prepare_cached(&sql)?;
